requirements.txt
asgiref==3.9.1
django==5.2.6
django-environ==0.12.0
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
sqlparse==0.5.3
